Stability fixes (#511)
* remove EmbedUrl from live now query and remove old live modal

* Add ability to open featured cards in discover tab in-app
